+++ packages/components/button/src/Button.stories.tsx
@@ -1,8 +1,9 @@
+import { Checkbox } from '@spark-ui/checkbox'
 import { Icon } from '@spark-ui/icon'
 import { Check } from '@spark-ui/icons/dist/icons/Check'
 import { FavoriteOutline } from '@spark-ui/icons/dist/icons/FavoriteOutline'
 import { Meta, StoryFn } from '@storybook/react'
-import { type ComponentProps } from 'react'
+import { type ComponentProps, useState } from 'react'
 
 import { Button } from '.'
 
@@ -32,7 +33,7 @@ const shapes: ButtonProps['shape'][] = ['rounded', 'square', 'pill']
 export const Default: StoryFn = _args => <Button>Default button</Button>
 
 export const Sizes: StoryFn = _args => (
-  <div className="gap-md flex flex-wrap items-center">
+  <div className="flex flex-wrap items-center gap-md">
     {sizes.map(size => {
       return (
         <Button key={size} size={size}>
@@ -44,7 +45,7 @@ export const Sizes: StoryFn = _args => (
 )
 
 export const Shapes: StoryFn = _args => (
-  <div className="gap-md flex flex-wrap items-center">
+  <div className="flex flex-wrap items-center gap-md">
     {shapes.map(shape => {
       return (
         <Button key={shape} shape={shape}>
@@ -58,7 +59,7 @@ export const Shapes: StoryFn = _args => (
 export const Disabled: StoryFn = _args => <Button disabled>Disabled button</Button>
 
 export const Design: StoryFn = _args => (
-  <div className="gap-md flex flex-wrap">
+  <div className="flex flex-wrap gap-md">
     {designs.map(design => (
       <Button key={design} design={design}>
         {design} button
@@ -68,7 +69,7 @@ export const Design: StoryFn = _args => (
 )
 
 export const Intent: StoryFn = _args => (
-  <div className="gap-md flex flex-wrap">
+  <div className="flex flex-wrap gap-md">
     {intents.map(intent => (
       <Button key={intent} intent={intent}>
         {intent} button
@@ -78,7 +79,7 @@ export const Intent: StoryFn = _args => (
 )
 
 export const Icons: StoryFn = _args => (
-  <div className="gap-md flex flex-wrap">
+  <div className="flex flex-wrap gap-md">
     <Button>
       Button
       <Icon>
@@ -103,10 +104,59 @@ export const Icons: StoryFn = _args => (
   </div>
 )
 
+export const Loading: StoryFn = () => {
+  const [isLoading, setIsLoading] = useState(true)
+
+  return (
+    <div className="flex flex-col items-start gap-lg">
+      <Checkbox checked={isLoading} onClick={() => setIsLoading(!isLoading)}>
+        Toggle loading state
+      </Checkbox>
+
+      <Button isLoading={isLoading} loadingLabel="Loading...">
+        <Icon>
+          <FavoriteOutline />
+        </Icon>
+        Button (width is preserved)
+      </Button>
+    </div>
+  )
+}
+
+export const LoadingWithText: StoryFn = () => {
+  const [isLoading, setIsLoading] = useState(true)
+
+  return (
+    <div className="flex flex-col items-start gap-lg">
+      <Checkbox checked={isLoading} onClick={() => setIsLoading(!isLoading)}>
+        Toggle loading state
+      </Checkbox>
+
+      <Button isLoading={isLoading} loadingText="Loading...">
+        <Icon>
+          <FavoriteOutline />
+        </Icon>
+        Button with long text
+      </Button>
+
+      <Button
+        isLoading={isLoading}
+        loadingText="Loading (spinner to the right)..."
+        spinnerPlacement="right"
+      >
+        <Icon>
+          <FavoriteOutline />
+        </Icon>
+        Button with long text
+      </Button>
+    </div>
+  )
+}
+
 export const Link: StoryFn = _args => (
-  <div className="gap-md flex flex-wrap">
+  <div className="flex flex-wrap gap-md">
     <Button asChild>
-      <a href="/">button</a>
+      <a href="/">Button as a link</a>
     </Button>
   </div>
 )