+++ packages/components/popover/src/Popover.stories.tsx
@@ -38,10 +38,12 @@ export const Default: StoryFn = _args => {
         <Popover.Trigger asChild>
           <Button>Trigger popover</Button>
         </Popover.Trigger>
-        <Popover.Content>
-          some text
-          <Popover.Arrow />
-        </Popover.Content>
+        <Popover.Portal>
+          <Popover.Content>
+            Popover contents
+            <Popover.Arrow />
+          </Popover.Content>
+        </Popover.Portal>
       </Popover>
     </ShowcaseContainer>
   )
@@ -52,7 +54,7 @@ export const Controlled: StoryFn = () => {
 
   return (
     <>
-      <div className="gap-lg mb-lg flex">
+      <div className="mb-lg flex gap-lg">
         <Button onClick={() => setOpen(true)}>Open popover</Button>
         <Button onClick={() => setOpen(false)}>Close popover</Button>
       </div>
@@ -61,10 +63,12 @@ export const Controlled: StoryFn = () => {
           <Popover.Anchor asChild>
             <p>Popover is attached to this text (anchor)</p>
           </Popover.Anchor>
-          <Popover.Content onInteractOutside={() => setOpen(false)}>
-            some text
-            <Popover.Arrow />
-          </Popover.Content>
+          <Popover.Portal>
+            <Popover.Content onInteractOutside={() => setOpen(false)}>
+              Popover contents
+              <Popover.Arrow />
+            </Popover.Content>
+          </Popover.Portal>
         </Popover>
       </ShowcaseContainer>
     </>
@@ -75,7 +79,7 @@ export const Anchored: StoryFn = _args => {
   return (
     <ShowcaseContainer className="!justify-between">
       <Popover>
-        <div className="gap-md flex flex-col">
+        <div className="flex flex-col gap-md">
           <Popover.Trigger asChild>
             <Button>Trigger 1</Button>
           </Popover.Trigger>
@@ -90,10 +94,13 @@ export const Anchored: StoryFn = _args => {
         <Popover.Trigger asChild>
           <Button intent="secondary">Anchor element</Button>
         </Popover.Trigger>
-        <Popover.Content>
-          some text
-          <Popover.Arrow />
-        </Popover.Content>
+
+        <Popover.Portal>
+          <Popover.Content>
+            Popover contents
+            <Popover.Arrow />
+          </Popover.Content>
+        </Popover.Portal>
       </Popover>
     </ShowcaseContainer>
   )
@@ -106,11 +113,13 @@ export const MatchTriggerWidth: StoryFn = _args => {
         <Popover.Trigger asChild>
           <Button>Check the width of this popover</Button>
         </Popover.Trigger>
-        <Popover.Content matchTriggerWidth>
-          Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt
-          ut labore et dolore magna aliqua.
-          <Popover.Arrow />
-        </Popover.Content>
+        <Popover.Portal>
+          <Popover.Content matchTriggerWidth>
+            Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor
+            incididunt ut labore et dolore magna aliqua.
+            <Popover.Arrow />
+          </Popover.Content>
+        </Popover.Portal>
       </Popover>
     </ShowcaseContainer>
   )
@@ -120,16 +129,18 @@ export const Boundaries: StoryFn = () => {
   const [boundaryContainer, setBoundaryContainer] = useState<HTMLDivElement | null>(null)
 
   return (
-    <ShowcaseContainer ref={setBoundaryContainer} className="!w-sz-256 !h-sz-240">
+    <ShowcaseContainer ref={setBoundaryContainer} className="!h-sz-240 !w-sz-256">
       <Popover>
         <Popover.Trigger asChild>
           <Button>Trigger popover</Button>
         </Popover.Trigger>
-        <Popover.Content collisionBoundary={[boundaryContainer]}>
-          Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt
-          ut labore et dolore magna aliqua.
-          <Popover.Arrow />
-        </Popover.Content>
+        <Popover.Portal>
+          <Popover.Content collisionBoundary={[boundaryContainer]}>
+            Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor
+            incididunt ut labore et dolore magna aliqua.
+            <Popover.Arrow />
+          </Popover.Content>
+        </Popover.Portal>
       </Popover>
     </ShowcaseContainer>
   )
@@ -148,8 +159,8 @@ export const Positionning: StoryFn = _args => {
   }
 
   return (
-    <div className="gap-lg flex flex-col">
-      <div className="gap-lg flex">
+    <div className="flex flex-col gap-lg">
+      <div className="flex gap-lg">
         <p className="text-headline-2">Side:</p>
         <RadioGroup value={currentSide} onValueChange={handleChangeSide} orientation="horizontal">
           {['bottom', 'top', 'left', 'right'].map(side => (
@@ -160,7 +171,7 @@ export const Positionning: StoryFn = _args => {
         </RadioGroup>
       </div>
 
-      <div className="gap-lg flex">
+      <div className="flex gap-lg">
         <p className="text-headline-2">Align:</p>
         <RadioGroup value={currentAlign} onValueChange={handleChangeAlign} orientation="horizontal">
           {['start', 'center', 'end'].map(align => (
@@ -176,10 +187,12 @@ export const Positionning: StoryFn = _args => {
           <Popover.Trigger asChild>
             <Button>Trigger popover</Button>
           </Popover.Trigger>
-          <Popover.Content side={currentSide} align={currentAlign}>
-            some text
-            <Popover.Arrow />
-          </Popover.Content>
+          <Popover.Portal>
+            <Popover.Content side={currentSide} align={currentAlign}>
+              Popover contents
+              <Popover.Arrow />
+            </Popover.Content>
+          </Popover.Portal>
         </Popover>
       </ShowcaseContainer>
     </div>