+++ packages/components/radio-group/CHANGELOG.md
@@ -3,6 +3,34 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.8](https://github.com/adevinta/spark/compare/@spark-ui/radio-group@1.0.7...@spark-ui/radio-group@1.0.8) (2023-06-19)
+
+**Note:** Version bump only for package @spark-ui/radio-group
+
+## [1.0.7](https://github.com/adevinta/spark/compare/@spark-ui/radio-group@1.0.6...@spark-ui/radio-group@1.0.7) (2023-06-16)
+
+**Note:** Version bump only for package @spark-ui/radio-group
+
+## [1.0.6](https://github.com/adevinta/spark/compare/@spark-ui/radio-group@1.0.5...@spark-ui/radio-group@1.0.6) (2023-06-15)
+
+### Bug Fixes
+
+- undo radix label bump ([04b5d2f](https://github.com/adevinta/spark/commit/04b5d2f138d850f9d0bd14bc4e160299b46ac62b))
+
+## [1.0.5](https://github.com/adevinta/spark/compare/@spark-ui/radio-group@1.0.4...@spark-ui/radio-group@1.0.5) (2023-06-15)
+
+**Note:** Version bump only for package @spark-ui/radio-group
+
+## [1.0.4](https://github.com/adevinta/spark/compare/@spark-ui/radio-group@1.0.3...@spark-ui/radio-group@1.0.4) (2023-06-14)
+
+**Note:** Version bump only for package @spark-ui/radio-group
+
+## [1.0.3](https://github.com/adevinta/spark/compare/@spark-ui/radio-group@1.0.2...@spark-ui/radio-group@1.0.3) (2023-06-12)
+
+### Bug Fixes
+
+- **radio-group:** update intent success, alert and error styles ([a41e50c](https://github.com/adevinta/spark/commit/a41e50cc6f20d1e99431a195c7694ab34d375086))
+
 ## [1.0.2](https://github.com/adevinta/spark/compare/@spark-ui/radio-group@1.0.1...@spark-ui/radio-group@1.0.2) (2023-06-08)
 
 **Note:** Version bump only for package @spark-ui/radio-group