+++ packages/components/tabs/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.2.7](https://github.com/adevinta/spark/compare/@spark-ui/tabs@1.2.6...@spark-ui/tabs@1.2.7) (2023-06-19)
+
+**Note:** Version bump only for package @spark-ui/tabs
+
+## [1.2.6](https://github.com/adevinta/spark/compare/@spark-ui/tabs@1.2.5...@spark-ui/tabs@1.2.6) (2023-06-19)
+
+**Note:** Version bump only for package @spark-ui/tabs
+
+## [1.2.5](https://github.com/adevinta/spark/compare/@spark-ui/tabs@1.2.4...@spark-ui/tabs@1.2.5) (2023-06-16)
+
+**Note:** Version bump only for package @spark-ui/tabs
+
+## [1.2.4](https://github.com/adevinta/spark/compare/@spark-ui/tabs@1.2.3...@spark-ui/tabs@1.2.4) (2023-06-14)
+
+**Note:** Version bump only for package @spark-ui/tabs
+
 ## [1.2.3](https://github.com/adevinta/spark/compare/@spark-ui/tabs@1.2.2...@spark-ui/tabs@1.2.3) (2023-06-09)
 
 **Note:** Version bump only for package @spark-ui/tabs